Split UI validation
