+++ src/game_data_validation_internal.h
@@ -155,6 +155,10 @@ bool validate_render_effects(validation_context *ctx, yyjson_val *root, validati
 bool validate_lights(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool validate_transitions(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool validate_render_settings(validation_context *ctx, yyjson_val *root);
+bool ui_metric_name_valid(const char *metric);
+bool validate_ui_panels(validation_context *ctx, yyjson_val *panels, const char *path, validation_names *names);
+bool validate_ui_inspectors(validation_context *ctx, yyjson_val *inspectors, const char *path, validation_names *names);
+bool validate_ui(validation_context *ctx, yyjson_val *root, validation_names *names);
 bool is_single_byte_string(yyjson_val *value);
 
 bool is_vec_array(yyjson_val *value, size_t min_count);
